decomp: repair 28 unverified semantic drafts (+1 new function, 409/633 in src/matched)
Audit found 39 entries marked semantic whose C never compiled or never
matched retail; the ROM links asm/matchings/*.s, so make compare could not
see them. Fixed struct offsets (Unk0748, Unk37508Work), prototype/type
mismatches, a K&R vs variadic no-op, and a missing gData_08094BB4 symbol.
Verified semantic C: 365 -> 391.

New tooling: matched_rescore.py (audit linked drafts), fix_prototypes.py.

+++ src/matched/sub_08033E3C.c
@@ -10,7 +10,7 @@ void sub_08033E3C(struct Unk33F30 *a)
     if (v <= 0)
     {
         sub_08033D90(a);
-        a->unk2C = a->unk34.unk0C;
+        a->unk2C = (s32)a->unk34.unk0C;
     }
 }
