+++ src/matched/sub_08062A1C.c
@@ -6,7 +6,7 @@ void sub_08062A1C(u32 v)
     u32 tmp[2];
     u32 base;
 
-    tmp[0] = gUnk_030007B0;
+    tmp[0] = (u32)gUnk_030007B0;
     base = tmp[0];
     tmp[0] = 0x114;
     tmp[0] = base + tmp[0];
