+++ src/matched/sub_08070DF4.c
@@ -4,6 +4,6 @@
 u8 sub_08070DF4(struct Unk7069C *a, void *b, u8 c)
 {
     sub_0807069C(a);
-    return sub_08070D44(a, b, c);
+    return sub_08070D44(a, (s32)b, c);
 }
